Move base api related test classes in st2tests.api module.
This way they can be re-used outside the st2api scope.
