+++ st2stream/tests/unit/controllers/v1/test_stream.py
@@ -22,7 +22,8 @@ from st2common.persistence.action import Action, RunnerType
 from st2stream.controllers.v1 import stream
 import st2stream.listener
 from st2tests.api import SUPER_SECRET_PARAMETER
-from base import FunctionalTest
+from st2api.tests.base import FunctionalTest
+#from base import FunctionalTest
 
 
 RUNNER_TYPE_1 = {
